% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{gwr_diagnostic_extractors}
\alias{gwr_diagnostic_extractors}
\alias{gwr_rss}
\alias{gwr_aicc}
\alias{gwr_cv_score}
\title{Compact diagnostic extractors}
\usage{
gwr_rss(object, ...)

gwr_aicc(object, ...)

gwr_cv_score(object, ...)
}
\arguments{
\item{object}{A fitted `gwrs_fit` object.}

\item{...}{Unused.}
}
\value{
A single numeric diagnostic value.
}
\description{
Compact diagnostic extractors
}
\examples{
quake <- datasets::quakes[1:32, ]
coords <- as.matrix(quake[, c("long", "lat")])
x <- as.matrix(quake[, c("depth", "stations")])
fit <- gwr_fit(
  x, quake$mag, coords = coords, k = 20,
  control = gwrs_control(n_threads = 1)
)
c(rss = gwr_rss(fit), aicc = gwr_aicc(fit), cv = gwr_cv_score(fit))

}
