% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_models.R
\name{gwr_reference_formulas}
\alias{gwr_reference_formulas}
\alias{gwr_lasso}
\alias{gwr_ridge}
\alias{gwr_en}
\title{Formula interfaces for reference penalized GWR models}
\usage{
gwr_lasso(formula, data, coords, ...)

gwr_ridge(formula, data, coords, ...)

gwr_en(formula, data, coords, ..., alpha = 0.5)
}
\arguments{
\item{formula}{Model formula with an intercept.}

\item{data}{Data frame containing model variables.}

\item{coords}{Optional coordinate matrix used to construct neighbors.}

\item{...}{Additional arguments passed to [gwr_penalized()].}

\item{alpha}{Elastic-net mixing parameter in `[0, 1]`.}
}
\value{
An object inheriting from `gwrs_fit`.
}
\description{
These are concise formula wrappers around [gwr_penalized()].
}
\examples{
quake <- datasets::quakes[1:30, ]
control <- gwrs_control(n_threads = 1)
lasso <- gwr_lasso(
  mag ~ depth + stations, quake, c("long", "lat"),
  k = 18, lambda = 0.05, control = control
)
ridge <- gwr_ridge(
  mag ~ depth + stations, quake, c("long", "lat"),
  k = 18, lambda = 0.05, control = control
)
elastic_net <- gwr_en(
  mag ~ depth + stations, quake, c("long", "lat"),
  k = 18, lambda = 0.05, alpha = 0.6, control = control
)
c(lasso$method, ridge$method, elastic_net$method)

}
