% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_sl.R
\name{gwr_sl}
\alias{gwr_sl}
\title{Formula interface for geographically weighted scaled lasso}
\usage{
gwr_sl(formula, data, coords, ...)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame containing model variables.}

\item{coords}{Coordinate matrix or two column names in `data`.}

\item{...}{Additional arguments passed to [gwr_sl_fit()].}
}
\value{
An object of class `gwrs_fit`.
}
\description{
Formula interface for geographically weighted scaled lasso
}
\examples{
quake <- datasets::quakes[1:32, ]
fit <- gwr_sl(
  mag ~ depth + stations,
  data = quake,
  coords = c("long", "lat"),
  k = 20,
  lambda = 0.05,
  alpha = 0.7,
  d = 0.5,
  control = gwrs_control(n_threads = 1)
)
fit

}

\keyword{internal}
