% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gwrs_path}
\alias{plot.gwrs_path}
\title{Plot a gwrs regularization path}
\usage{
\method{plot}{gwrs_path}(
  x,
  type = c("criterion", "active", "convergence"),
  criterion = c("aicc", "bic", "gcv", "ebic", "ebic_refit"),
  ...
)
}
\arguments{
\item{x}{A `gwrs_path`.}

\item{type}{Plot an information criterion, mean active count, or convergence.}

\item{criterion}{Criterion column in `x$path_summary`.}

\item{...}{Additional arguments passed to [graphics::plot()].}
}
\value{
Invisibly returns `x`.
}
\description{
Plot a gwrs regularization path
}
\examples{
quake <- datasets::quakes[1:30, ]
coords <- as.matrix(quake[, c("long", "lat")])
x <- as.matrix(quake[, c("depth", "stations")])
path <- gwr_lasso_path(
  x, quake$mag, coords = coords, k = 18,
  lambda = c(0.2, 0.08, 0.02),
  control = gwrs_control(n_threads = 1, diagnostics = "none")
)
plot(path, type = "active")

}
