% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gwrs_fit}
\alias{summary.gwrs_fit}
\title{Summarize a fitted gwrs model}
\usage{
\method{summary}{gwrs_fit}(object, ...)
}
\arguments{
\item{object}{A fitted `gwrs_fit` object.}

\item{...}{Unused.}
}
\value{
An object of class `summary.gwrs_fit`.
}
\description{
Summarize a fitted gwrs model
}
\examples{
quake <- datasets::quakes[1:32, ]
coords <- as.matrix(quake[, c("long", "lat")])
x <- as.matrix(quake[, c("depth", "stations")])
fit <- gwr_lasso_fit(
  x, quake$mag, coords = coords, k = 20, lambda = 0.05,
  control = gwrs_control(n_threads = 1)
)
summary(fit)

}
